+++ shardingsphere-sql-parser/shardingsphere-sql-parser-dialect/shardingsphere-sql-parser-postgresql/src/main/antlr4/imports/postgresql/DALStatement.g4
@@ -63,7 +63,7 @@ explain
     ;
 
 explainableStmt
-    : select | insert | update | delete | declare | execute | createMaterializedView | refreshMatViewStmt
+    : select | insert | update | delete | declare | executeStmt | createMaterializedView | refreshMatViewStmt
     ;
 
 explainOptionList
