+++ shardingsphere-sql-parser/shardingsphere-sql-parser-dialect/shardingsphere-sql-parser-postgresql/src/main/antlr4/imports/postgresql/DMLStatement.g4
@@ -521,10 +521,6 @@ copyDelimiter
     : (USING)? DELIMITERS STRING_
     ;
 
-deallocate
-    : DEALLOCATE PREPARE? (name | ALL)
-    ;
-
 fetch
     : FETCH fetchArgs
     ;
