+++ shardingsphere-sql-parser/shardingsphere-sql-parser-dialect/shardingsphere-sql-parser-postgresql/src/main/antlr4/org/apache/shardingsphere/sql/parser/autogen/PostgreSQLStatement.g4
@@ -68,5 +68,8 @@ execute
     | dropView
     | dropSequence
     | vacuum
+    | prepare
+    | executeStmt
+    | deallocate
     ) SEMI_?
     ;
