+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/main/java/org/apache/shardingsphere/sql/parser/core/visitor/SQLVisitorRule.java
@@ -112,8 +112,14 @@ public enum SQLVisitorRule {
 
     DROP_SEQUENCE("DropSequence", SQLStatementType.DDL),
     
-    ALTER_SYNONYM("AlterSynonym", SQLStatementType.DDL),  
-  
+    ALTER_SYNONYM("AlterSynonym", SQLStatementType.DDL),
+    
+    PREPARE("Prepare", SQLStatementType.DDL),
+    
+    EXECUTE_STMT("ExecuteStmt", SQLStatementType.DDL),
+    
+    DEALLOCATE("Deallocate", SQLStatementType.DDL),
+    
     SET_TRANSACTION("SetTransaction", SQLStatementType.TCL),
     
     SET_IMPLICIT_TRANSACTIONS("SetImplicitTransactions", SQLStatementType.TCL),
